perf(tests): extract _app.getInitialProps into its own module
`app-settings-bootstrap.test.ts` pins the auth-cookie discriminator in
`_app.getInitialProps` and had to import `~/pages/_app` to reach it, which loaded
the whole provider and layout tree - 285 modules that entered the unit suite's
module union for that one test file.

The body moves to `~/utils/app-initial-props`; `_app.tsx` assigns it, so it still
ships in the same server and client bundles and still runs on client navigations.
